Clean up code structure
- Separate methods into a separate file (modules/game_actions.rb) and
organize them into a module (GameActions)
- WIN_COMBOS constant should also be in this module
- Player class should be in a separate class -> scalability purposes -
when we want to later on have a human vs computer game
- Require player.rb and game_actions.rb in game.rb and include
GameActions inside class Game
