+++ classes/player.rb
@@ -1,0 +1,9 @@
+# Player classes that will be playing the game
+class Player
+  attr_accessor :name, :sign, :moves
+  def initialize(name, sign)
+    @name = name
+    @sign = sign
+    @moves = []
+  end
+end
