+++ main.rb
@@ -1,3 +1,3 @@
-require_relative 'classes/game'
+require_relative "classes/game"
 demo = Game.new
 demo.game_start
